# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Lamontville Golden Arrows FC API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List squad players
      parameters:
        - in: query
          name: position
          schema: { type: string }
      responses:
        "200":
          description: Players list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create player (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player profile
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    put:
      operationId: updatePlayer
      tags: [players]
      summary: Update player (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted

  # ── STAFF ─────────────────────────────────────────────────────────────────
...
  /results:
    get:
      operationId: listResults
      tags: [results]
      summary: List recent match results
      parameters:
        - in: query
          name: limit
          schema: { type: integer }
      responses:
        "200":
          description: Results list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchResult"
  /results/{id}:
    get:
      operationId: getResult
      tags: [results]
      summary: Get match result details
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Match result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResult"

  # ── LEAGUE TABLE ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Player:
      type: object
      required: [id, name, position, number, nationality]
      properties:
        id: { type: integer }
        name: { type: string }
        position: { type: string }
        number: { type: integer }
        nationality: { type: string }
        age: { type: ["integer", "null"] }
        photoUrl: { type: ["string", "null"] }
        bio: { type: ["string", "null"] }
        appearances: { type: integer }
        goals: { type: integer }
        assists: { type: integer }
    PlayerInput:
      type: object
      required: [name, position, number, nationality]
      properties:
        name: { type: string }
        position: { type: string }
        number: { type: integer }
        nationality: { type: string }
        age: { type: integer }
        photoUrl: { type: string }
        bio: { type: string }
        appearances: { type: integer }
        goals: { type: integer }
        assists: { type: integer }

    # ── STAFF ─────────────────────────────────────────────────────────────
...
    MatchResult:
      type: object
      required: [id, date, homeTeam, awayTeam, homeScore, awayScore, competition]
      properties:
        id: { type: integer }
        date: { type: string }
        homeTeam: { type: string }
        awayTeam: { type: string }
        homeScore: { type: integer }
        awayScore: { type: integer }
        competition: { type: string }
        venue: { type: ["string", "null"] }
        scorers: { type: array, items: { type: string } }
        matchReport: { type: ["string", "null"] }
        highlightUrl: { type: ["string", "null"] }

    # ── LEAGUE TABLE ──────────────────────────────────────────────────────
...