# lib/api-zod/src/generated/types/player.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamontville Golden Arrows FC API
 * OpenAPI spec version: 0.1.0
 */

export interface Player {
  id: number;
  name: string;
  position: string;
  number: number;
  nationality: string;
  /** @nullable */
  age?: number | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  appearances?: number;
  goals?: number;
  assists?: number;
}