# lib/api-zod/src/generated/types/playerInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Lamontville Golden Arrows FC API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerInput {
  name: string;
  position: string;
  number: number;
  nationality: string;
  age?: number;
  photoUrl?: string;
  bio?: string;
  appearances?: number;
  goals?: number;
  assists?: number;
}